requirements/requirements-all.txt
logmuse
fastembed
peppy>=0.40.7
python-dotenv
qdrant-client>=1.16.1
psycopg>=3.3.2
ubiquerg
tqdm
flatdict>=4.0.1
pepdbagent>=0.12.3
